chore: include Node releases data in bundle
